Add Openshift 4 controller manager config and deployment
